Adds ability to overwrite config file
The e2e tests with minikube and AWS S3 as a
backup storage were failing. This was mainly
because the cli was not able to overwrite the
content of an already existing backup-store-config.yaml
file. This is needed, otherwise the config file would
keep the local minio address for the backup
storage and try to connect tho minio instead of AWS S3.
The changes here enable an overwrite of the config file.